rm ZomiAttack
