+++ core/src/ecs/entities/characters/Piggo.ts
@@ -55,7 +55,7 @@ const npcOnTick = (entity: Entity<Position>, world: World): void | InvokedAction
   const edibles = world.queryEntities(["food", "position"])
     .filter((e) => !(e.id.includes("piggo"))) as Entity<Food | Position>[]
 
-  const closest = closestEntity(edibles, position.data, 250)
+  const closest = closestEntity(edibles, position.data, 200)
 
   if (closest) return { action: "chase", params: { target: closest } }
