+++ core/src/ecs/entities/characters/Zomi.ts
@@ -1,7 +1,8 @@
 import {
   Actions, Chase, Collider, Debug, Entity, Health, InvokedAction,
-  NPC, Networked, Position, PositionProps, Renderable, World, ZomiAttack,
-  closestEntity, positionDelta, loadTexture, round, randomInt, max, Element
+  NPC, Networked, Position, PositionProps, Renderable, World,
+  closestEntity, positionDelta, loadTexture, round, randomInt, max, Element,
+  Action, ZomiAttackSounds, playSound, randomChoice
 } from "@piggo-gg/core"
 import { AnimatedSprite } from "pixi.js"
 
@@ -75,3 +76,14 @@ const npcOnTick = (entity: Entity<Position>, world: World): void | InvokedAction
 
   return { action: "chase", params: { target: closest } }
 }
+
+export const ZomiAttack = (damage: number, cooldown: number) => Action<{ target: Entity }>(({ entity, params, world }) => {
+  const { target } = params
+  const { health } = target.components
+
+  if (health) health.data.health -= damage
+
+  entity?.components.position?.clearHeading()
+
+  playSound(world.client?.sounds[randomChoice(["attack1", "attack2", "attack3", "attack4"]) as ZomiAttackSounds])
+}, cooldown)
